refactor: 회원가입 테스트 추가
