+++ src/main/java/controller/HandlerMapper.java
@@ -14,6 +14,7 @@ import java.util.Objects;
 import static response.ResponseBody.responseBody;
 import static response.ResponseHeader.response200Header;
 import static utils.FileIoUtils.*;
+import static utils.IOUtils.*;
 
 public class HandlerMapper {
     private final BufferedReader br;
